Add components needed for blog
Adjustment to config file for a slightly cleaner URL scheme.
This is not final (see #4).

Add two new template files. One for the blog index page and the
other for the post content page. The structure looks ok but the
styling needs some work. That will have to come later.

Signed-off-by: Amir Chaudhry <amir@docker.com>